refactor: database->repository 변경
