+++ src/main/java/roomescape/repositiory/ReservationRepositoryImpl.java
@@ -1,10 +1,10 @@
-package roomescape.database;
+package roomescape.repositiory;
 
 import java.util.ArrayList;
 import java.util.List;
 import roomescape.domain.Reservation;
 
-public class ReservationDatabaseImpl implements ReservationDatabase {
+public class ReservationRepositoryImpl implements ReservationRepository {
 
     private final List<Reservation> reservations = new ArrayList<>();
