+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,32 +2,32 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.database.ReservationDatabase;
-import roomescape.database.ReservationDatabaseImpl;
+import roomescape.repositiory.ReservationRepository;
+import roomescape.repositiory.ReservationRepositoryImpl;
 import roomescape.domain.Reservation;
 import roomescape.domain.dto.ReservationRequestDto;
 
 @Service
 public class ReservationService {
 
-    private final ReservationDatabase reservationDatabase;
+    private final ReservationRepository reservationRepository;
 
     public ReservationService() {
-        this.reservationDatabase = new ReservationDatabaseImpl();
+        this.reservationRepository = new ReservationRepositoryImpl();
     }
 
     public List<Reservation> readAll() {
-        return reservationDatabase.findAll();
+        return reservationRepository.findAll();
     }
 
     public Reservation add(ReservationRequestDto reservationDto) {
         Reservation reservation = convertReservation(reservationDto);
-        return reservationDatabase.add(reservation);
+        return reservationRepository.add(reservation);
     }
 
     public void delete(Long id) {
-        Reservation reservation = reservationDatabase.findById(id);
-        reservationDatabase.delete(reservation);
+        Reservation reservation = reservationRepository.findById(id);
+        reservationRepository.delete(reservation);
     }
 
     private Reservation convertReservation(ReservationRequestDto reservationDto) {
